+++ app/preferences-renderer.js
@@ -92,6 +92,7 @@ window.onload = async (e) => {
         document.querySelector('#electron').innerHTML = await window.runtime.electron()
         document.querySelector('#platform').innerHTML = await window.runtime.platform()
         document.querySelector('#windowsStore').innerHTML = await window.runtime.windowsStore() || false
+        document.querySelector('#windowsPortable').innerHTML = await window.runtime.windowsPortable() || false
       }
       setWindowHeight()
     }